requirements.txt
Django==2.2.7
djangorestframework==3.10.3
PyJWT==1.7.1
pytz==2019.3

mysqlclient==1.4.6

django-cors-headers==3.2.0